only finish answer and permission

+++ src/main/java/com/example/demo/service/QuestionService.java
@@ -5,5 +5,5 @@ import com.example.demo.model.Permission;
 public interface QuestionService {
     Boolean createQuestion(String statement, Permission permission);
     Boolean editQuestion(Integer questionid, Permission permission, String newStatement);
-    Boolean deleteQuestion(Integer questionid, Permission permission);
+    Boolean deleteQuestion(Long questionid, Permission permission);
 }
